website-wizard/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication / Password Hashing

bcrypt==5.0.0

python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Validation / Configuration

pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

python-dotenv==1.0.1
requests==2.31.0

# AI / Analysis

openai==1.17.0
beautifulsoup4==4.12.3

# Utilities

tqdm==4.66.2
